Convert binaries to benchmarks
